backend/requirements.txt
fastapi 
uvicorn
psycopg2-binary 
sqlalchemy 
pydantic 
email-validator